Some sort of basic module registering for main menu building.
I'm not satisfied yet. But basically you register your views as a
menu_item and a context processor will provide a base_menu used to build
the menu un the base template. I recommend registering views for menu in
the urls.py of the apps. see example in core.urls