Keep it simple. Reword the menu
In fact this kind of simple menu will be sufficient as we will most
certainly add only one entry to per app in the main menu which redirect
to a given index of the app. The app then can do whatever it wants. I
plan to create a submenu "wrapper" with tags to create menu item in an
unified way so each app can have a custom submenu without bkreaking the
design if they extends base.

Of course login/logout will not stay in the current menu and will have a
special treatment, so for the moment the dual selected menu is ok.