Carshare module.
It's kind of working but it would be mush more dynamic if the
accept/refuse buttons were done in javascript. Also currently the
carshare description does not contains all interresting data.  Finally
it will be way better when i will add a section for people who are
looking for a car share to post a request. and with propoer table style.