Update shop
It is now possible to create pack of products. When a pack contain
product with actions they will all be executed sequencialy with no
priority managment. Buying history has been refactored and now it
contains a reference to the product/pack so we will be able to fetch
detailled informations on the product/pack. It's a bit dirty and futher
modification of the product administration should be done while keeping
in mind you should never really delete a product/pack but instead
disable it. The same applies to a future product/pack edition which will
require to create a new product/pack and disable the old one. If you do
not respect this it will **break the shop history** by displaying modified
product/pack information from what was originally bought.

Next update will give user more details about the product in it's
history list.